+++ CHANGELOG.md
@@ -14,7 +14,7 @@ and this project adheres to
 - ✨(back) add celery for running background tasks
 - 🧱(helm) add celery worker and beat deployments
 - ✨(front) add "Edit in Docs" action to export a message to Docs
-
+- ✨(back) add output token limit per message
 
 ### Changed